refactor: move handlers (prev. api) and parser from pkg to internal
